break reusable sections into separate components
